allows looping for multiple purchases

+++ managePurchases.py
@@ -266,31 +266,36 @@ if __name__ == "__main__":
         db.connect(config.db_config)
         setMetals(db) # Sets value of data.metals for translation when making CoinData objects
 
-        if args["delete"]: # Delete mode
-            coin = getCoinInformation(db,{"show_only_owned":True})
-            purchase = selectPurchase(db,coin)
-            if alterDatabaseConfirmation():
-                result = db.deleteById({"purchases":purchase[0]})
-                printResult(result,purchase[2])
-                # Check if specific_coin is used by other purchases
-                if not db.fetchPurchasesWithSpecificCoinId(purchase[1]):
-                    # It is not used, see if user wants to delete it
-                    if getConfirmation("No remaining coins use the specific_coin information of the deleted coin. Delete entry from specific_coins table?"):
-                        result = db.deleteById({"specific_coins":purchase[1]})
-                        printResult(result,purchase[1])
-                    else:
-                        print("Keeping specific coin information")
-            else: # User didn't want to alter database
-                print("Aborting...")
-        else: # Default add mode
-            coin = getCoinInformation(db)
-            purchase = getPurchaseInformation()
-            specific_coin = getSpecificCoinInformation()
-            if alterDatabaseConfirmation():
-                specific_coin_id = pushSpecificCoin(db,specific_coin)
-                pushPurchase(db,coin,purchase,specific_coin_id)
-            else:
-                print("Aborting...")
+        while True:
+            if args["delete"]: # Delete mode
+                coin = getCoinInformation(db,{"show_only_owned":True})
+                purchase = selectPurchase(db,coin)
+                if alterDatabaseConfirmation():
+                    result = db.deleteById({"purchases":purchase[0]})
+                    printResult(result,purchase[2])
+                    # Check if specific_coin is used by other purchases
+                    if purchase[1] is not None and not db.fetchPurchasesWithSpecificCoinId(purchase[1]):
+                        # It is not used, see if user wants to delete it
+                        if getConfirmation("No remaining coins use the specific_coin information of the deleted coin. Delete entry from specific_coins table?"):
+                            result = db.deleteById({"specific_coins":purchase[1]})
+                            printResult(result,purchase[1])
+                        else:
+                            print("Keeping specific coin information")
+                else: # User didn't want to alter database
+                    print("Aborting...")
+                    exit(0)
+            else: # Default add mode
+                coin = getCoinInformation(db)
+                purchase = getPurchaseInformation()
+                specific_coin = getSpecificCoinInformation()
+                if alterDatabaseConfirmation():
+                    specific_coin_id = pushSpecificCoin(db,specific_coin)
+                    pushPurchase(db,coin,purchase,specific_coin_id)
+                else:
+                    print("Aborting...")
+                    exit(0)
+            if not getConfirmation(f"{'Delete' if args['delete'] else 'Add'} another purchase?"):
+                break
 
     finally:
         db.closeConnection()
